Add dynamic spawning of agents + doc refactor
